+++ sdk/metric/pipeline_test.go
@@ -25,7 +25,10 @@ import (
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/metric/unit"
 	"go.opentelemetry.io/otel/sdk/instrumentation"
+	"go.opentelemetry.io/otel/sdk/metric/aggregation"
 	"go.opentelemetry.io/otel/sdk/metric/metricdata"
+	"go.opentelemetry.io/otel/sdk/metric/metricdata/metricdatatest"
+	"go.opentelemetry.io/otel/sdk/metric/view"
 	"go.opentelemetry.io/otel/sdk/resource"
 )
 
@@ -61,7 +64,7 @@ func TestEmptyPipeline(t *testing.T) {
 }
 
 func TestNewPipeline(t *testing.T) {
-	pipe := newPipeline(nil)
+	pipe := newPipeline(nil, nil, nil)
 
 	output, err := pipe.produce(context.Background())
 	require.NoError(t, err)
@@ -158,7 +161,7 @@ func TestPipelineDuplicateRegistration(t *testing.T) {
 	}
 	for _, tt := range testCases {
 		t.Run(tt.name, func(t *testing.T) {
-			pipe := newPipeline(nil)
+			pipe := newPipeline(nil, nil, nil)
 			err := pipe.addAggregator(instrumentation.Scope{}, "name", "desc", unit.Dimensionless, testSumAggregator{})
 			require.NoError(t, err)
 
@@ -177,7 +180,7 @@ func TestPipelineDuplicateRegistration(t *testing.T) {
 
 func TestPipelineUsesResource(t *testing.T) {
 	res := resource.NewWithAttributes("noSchema", attribute.String("test", "resource"))
-	pipe := newPipeline(res)
+	pipe := newPipeline(res, nil, nil)
 
 	output, err := pipe.produce(context.Background())
 	assert.NoError(t, err)
@@ -185,7 +188,7 @@ func TestPipelineUsesResource(t *testing.T) {
 }
 
 func TestPipelineConcurrency(t *testing.T) {
-	pipe := newPipeline(nil)
+	pipe := newPipeline(nil, nil, nil)
 	ctx := context.Background()
 
 	var wg sync.WaitGroup
@@ -210,4 +213,62 @@ func TestPipelineConcurrency(t *testing.T) {
 		}()
 	}
 	wg.Wait()
+}
+
+func TestDefaultViewImplicit(t *testing.T) {
+	t.Run("Int64", testDefaultViewImplicit[int64]())
+	t.Run("Float64", testDefaultViewImplicit[float64]())
+}
+
+func testDefaultViewImplicit[N int64 | float64]() func(t *testing.T) {
+	inst := view.Instrument{
+		Scope:       instrumentation.Scope{Name: "testing/lib"},
+		Name:        "requests",
+		Description: "count of requests received",
+		Kind:        view.SyncCounter,
+		Aggregation: aggregation.Sum{},
+	}
+	return func(t *testing.T) {
+		reader := NewManualReader()
+		v, err := view.New(view.MatchInstrumentName("foo"), view.WithRename("bar"))
+		require.NoError(t, err)
+
+		tests := []struct {
+			name string
+			pipe *pipeline
+		}{
+			{
+				name: "NoView",
+				pipe: newPipeline(nil, reader, nil),
+			},
+			{
+				name: "NoMatchingView",
+				pipe: newPipeline(nil, reader, []view.View{v}),
+			},
+		}
+
+		for _, test := range tests {
+			t.Run(test.name, func(t *testing.T) {
+				i := newInserter[N](test.pipe)
+				got, err := i.Instrument(inst, unit.Dimensionless)
+				require.NoError(t, err)
+				assert.Len(t, got, 1, "default view not applied")
+
+				out, err := test.pipe.produce(context.Background())
+				require.NoError(t, err)
+				require.Len(t, out.ScopeMetrics, 1, "Aggregator not registered with pipeline")
+				sm := out.ScopeMetrics[0]
+				require.Len(t, sm.Metrics, 1, "metrics not produced from default view")
+				metricdatatest.AssertEqual(t, metricdata.Metrics{
+					Name:        inst.Name,
+					Description: inst.Description,
+					Unit:        unit.Dimensionless,
+					Data: metricdata.Sum[N]{
+						Temporality: metricdata.CumulativeTemporality,
+						IsMonotonic: true,
+					},
+				}, sm.Metrics[0], metricdatatest.IgnoreTimestamp())
+			})
+		}
+	}
 }